master

+++ application/controllers/Client.php
@@ -15,7 +15,12 @@
 
 		
 		}
-
+	public function get_all_user()
+	{
+		$email = $this->input->post('email');
+		$result = $this->User_model->get_all();
+		echo json_encode($result);
+	}
 
     	public function index() {
         	
@@ -161,9 +166,12 @@ public function client_more() {
         //print_r(array_values($data));
 
         
-        $this->Client_model->add($data2);
+       $this->Client_model->add($data2);
+        	redirect('/Client') ;
+        
+       
         //Loading View
-        redirect('/Client') ;
+        
 
         
         
@@ -232,7 +240,14 @@ public function client_more() {
         
         }
         else
-        {
+        {	
+        $this->session->set_flashdata('client_name',$this->input->post('name'));
+        $this->session->set_flashdata('address',$this->input->post('address'));
+        $this->session->set_flashdata('owner_name',$this->input->post('owner_name'));
+        $this->session->set_flashdata('category',$this->input->post('category'));
+        $this->session->set_flashdata('password',$this->input->post('password'));
+        $this->session->set_flashdata('img','abbas.png');
+        
         	echo '<script>alert("Same Client Already Exist");  window.location.href = window.location.origin+"/index.php/client/add_client";  </script>';
           	//redirect('Client/add_client') ;
         }
@@ -317,20 +332,73 @@ if($this->session->userdata('username')  && $this->session->userdata('type') == 
         $user_id = urldecode($this->uri->segment(3));
         $client_id = urldecode($this->uri->segment(4));
         $data = $this->Client_model->searchnm($client_id);
+        $assignmentData = $this->Client_model->searchnmAssignment($client_id);
+        $data = $this->User_model->get_all();
+        $isFlushable = true;
+        //$noOfAssignment = 0;
+        //print_r($assignmentData);
+       // foreach ($data as $key => $value1)
+  //       {
+         
+         $noOfAssignment = $value1['takenAssignment'];
+        foreach($assignmentData as $key => $value){
+        
+        $assignmentIds = $value['mysteryShopper_assignment_id'];
+        $mysteryShopper_id = $value['mysteryShopper_id'];
+       
+        
+        if($value['mysteryShopper_assignment_status']=='Working' || $value['mysteryShopper_assignment_status']=='Completed' || $value['mysteryShopper_assignment_status']=='Accept'){
+        $isFlushable=false;
+       
+        //exit;
+      
+    
+    
+        }else{
+        
+       
+      //  if($value1['user_id'] == $mysteryShopper_id && $value1['user_type'] =='MYSTERYSHOPPER'  )
+        //  {
+            //  $noOfAssignment-=1;
+              //echo $value1['user_id'] ;
+             // $takenAssignment = array('takenAssignment' => $noOfAssignment,);
+        //      $this->User_model->update($mysteryShopper_id,$takenAssignment);	
+        // }
+         
+         //$isFlushable=true;
+        
+        
+        }
+    }
+    
+    if ($isFlushable==true){
+    $this->Assignment_model->del($assignmentIds);
         unlink("uploads/".$data->mysteryShopper_client_image);
         $this->Client_model->del($user_id,$client_id);
         $this->Client_model->delClientinUserTable($user_id);
-        $status = 'Close';
-
-          $statusReject = array(
-          'mysteryShopper_assignment_status' => $status,
-           );
-
-         $this->Assignment_model->updateStatus($client_id,$statusReject);
-        redirect('/Client') ;
+         redirect('/Client') ;
+    }else {
+    $this->session->set_flashdata('status','Sorry you can not delete this client because he has assignment that is takken by MyStery Shopper');
+    redirect('/Client') ;
+    }
+        
+  //    }  
+      
+      if($assignmentData==null){
+      
+       unlink("uploads/".$data->mysteryShopper_client_image);
+        $this->Client_model->del($user_id,$client_id);
+        $this->Client_model->delClientinUserTable($user_id);
+         redirect('/Client') ;
+      
+      
+      }
+       
+       
+      
         }else
             {
-             $this->load->view("login");
+            // $this->load->view("login");
             }
